Agent it metrics (#9172)
* agent metrics  IT

* agent metrics  IT

* agent metrics  IT

* agent metrics  IT

* agent metrics  IT

* agent metrics  IT
